+++ tests/backward_compatibility/tabor_backward_compatibility_tests.py
@@ -5,7 +5,7 @@ import typing
 import importlib.util
 import sys
 
-from tests.hardware.old_tabor_simulator_based_tests import TaborSimulatorManager
+from tests.hardware.tabor_simulator_based_tests import TaborSimulatorManager
 from tests.hardware.dummy_devices import DummyDAC
 
 from qupulse.serialization import Serializer, FilesystemBackend, PulseStorage
